misc: split test files
